Transfer common logic to helper
